+++ clearwater_config_manager/scripts/config_validation/default_ifcs_schema.xsd
@@ -154,7 +154,7 @@
   </xs:complexType>
   <xs:complexType name="tInitialFilterCriteria">
     <xs:sequence>
-      <xs:element name="Priority" type="tPriority"/>
+      <xs:element name="Priority" type="tPriority" minOccurs="0"/>
       <xs:element name="TriggerPoint" type="tTrigger" minOccurs="0"/>
       <xs:element name="ApplicationServer" type="tApplicationServer"/>
       <xs:element name="ProfilePartIndicator" type="tProfilePartIndicator" minOccurs="0"/>
