--- clearwater_config_manager/scripts/remove_default_ifcs_config
@@ -1,57 +1,0 @@
-#!/bin/bash
-
-# Project Clearwater - IMS in the Cloud
-# Copyright (C) 2017  Metaswitch Networks Ltd
-#
-# This program is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or (at your
-# option) any later version, along with the "Special Exception" for use of
-# the program along with SSL, set forth below. This program is distributed
-# in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-# without even the implied warranty of MERCHANTABILITY or FITNESS FOR
-# A PARTICULAR PURPOSE.  See the GNU General Public License for more
-# details. You should have received a copy of the GNU General Public
-# License along with this program.  If not, see
-# <http://www.gnu.org/licenses/>.
-#
-# The author can be reached by email at clearwater@metaswitch.com or by
-# post at Metaswitch Networks Ltd, 100 Church St, Enfield EN2 6BQ, UK
-#
-# Special Exception
-# Metaswitch Networks Ltd  grants you permission to copy, modify,
-# propagate, and distribute a work formed by combining OpenSSL with The
-# Software, or a work derivative of such a combination, even if such
-# copying, modification, propagation, or distribution would otherwise
-# violate the terms of the GPL. You must comply with the GPL in all
-# respects for all of the code used other than OpenSSL.
-# "OpenSSL" means OpenSSL toolkit software distributed by the OpenSSL
-# Project and licensed under the OpenSSL Licenses, or a work based on such
-# software and licensed under the OpenSSL Licenses.
-# "OpenSSL Licenses" means the OpenSSL License and Original SSLeay License
-# under which the OpenSSL Project distributes the OpenSSL toolkit software,
-# as those licenses appear in the file LICENSE-OPENSSL.
-
-local_site_name=site1
-etcd_key=clearwater
-. /etc/clearwater/config
-. /usr/share/clearwater/utils/check-root-permissions 1
-
-# Check we can contact `etcd`
-if ! nc -z ${management_local_ip:-$local_ip} 4000
-then
-  echo "The Clearwater Configuration store (etcd) is not running"
-  echo "Please start it before removing configuration"
-  exit 2
-fi
-
-clearwater-etcdctl set /$etcd_key/$local_site_name/configuration/default_ifcs "<?xml version="1.0" encoding="UTF-8"?><DefaultIFCsSet></DefaultIFCsSet>" 2> /tmp/remove-default-ifcs-config.stderr.$$ > /tmp/remove-default-ifcs-config.stdout.$$
-rc=$?
-
-# Check the return code and log if appropriate.
-if [ $rc != 0 ] ; then
-  echo Unable to remove the Default IFCs                    >&2
-  cat /tmp/remove-default-ifcs-config.stderr.$$             >&2
-  cat /tmp/remove-default-ifcs-config.stdout.$$             >&2
-fi
-rm -f /tmp/remove-default-ifcs-config.stderr.$$ /tmp/remove-default-ifcs-config.stdout.$$
