+++ clearwater_config_manager/scripts/upload_default_ifcs_xml
@@ -1,0 +1,64 @@
+#!/bin/bash
+
+# Copyright (C) Metaswitch Networks 2017
+# If license terms are provided to you in a COPYING file in the root directory
+# of the source code repository by which you are accessing this code, then
+# the license outlined in that COPYING file applies to your use.
+# Otherwise no rights are granted except for those provided to you by
+# Metaswitch Networks in a separate written agreement.
+
+FILENAME=/etc/clearwater/default_ifcs.xml
+SCHEMA=/usr/share/clearwater/clearwater-config-manager/scripts/config_validation/default_ifcs_schema.xsd
+local_site_name=site1
+etcd_key=clearwater
+. /etc/clearwater/config
+
+. /usr/share/clearwater/utils/check-root-permissions 1
+
+# Check we can contact `etcd`
+if ! nc -z ${management_local_ip:-$local_ip} 4000
+then
+  echo "The Clearwater Configuration store (etcd) is not running"
+  echo "Please start it before uploading configuration"
+  exit 2
+fi
+
+# Check that the file is already present. The config file should either be put
+# in place by the user, or created by the plugin at startup if it finds both
+# the file and etcd key missing. If it does not exist, the disk may have been
+# full, or some other error may have occured.
+if [[ ! -f $FILENAME ]]
+then
+    echo "No configuration file at $FILENAME, unable to upload"
+    exit 2
+fi
+
+xmllint --format --pretty 1 --load-trace --debug --schema $SCHEMA $FILENAME 2> /tmp/upload-default-ifcs-xml.stderr.$$ > /tmp/upload-default-ifcs-xml.stdout.$$
+rc=$?
+
+# Check the return code and log if appropriate.
+if [ $rc != 0 ] ; then
+  echo The default IFC file is invalid - unable to upload >&2
+  cat /tmp/upload-default-ifcs-xml.stderr.$$              >&2
+  rm -f /tmp/upload-default-ifcs-xml.stderr.$$ /tmp/upload-default-ifcs-xml.stdout.$$
+  exit $rc
+fi
+rm -f /tmp/upload-default-ifcs-xml.stderr.$$ /tmp/upload-default-ifcs-xml.stdout.$$
+
+# Upload the file to etcd
+keypath=http://${management_local_ip:-$local_ip}:4000/v2/keys/$etcd_key/$local_site_name/configuration/default_ifcs
+curl -X PUT $keypath --data-urlencode value@$FILENAME 2> /tmp/upload-default-ifcs-xml.stderr.$$ | tee /tmp/upload-default-ifcs-xml.stdout.$$ | egrep -q "\"action\":\"set\""
+rc=$?
+
+# Check the return code and log if appropriate.
+if [ $rc != 0 ] ; then
+  echo Upload default IFCs configuration failed to $keypath  >&2
+  cat /tmp/upload-default-ifcs-xml.stderr.$$                 >&2
+  cat /tmp/upload-default-ifcs-xml.stdout.$$                 >&2
+fi
+rm -f /tmp/upload-default-ifcs-xml.stderr.$$ /tmp/upload-default-ifcs-xml.stdout.$$
+
+# Call the reload_default_ifc_xml scripts on this node
+/usr/share/clearwater/bin/reload_default_ifcs_xml
+
+exit $rc
