+++ clearwater_config_manager/scripts/validate_default_ifcs_xml
@@ -1,0 +1,38 @@
+#!/bin/bash
+
+# Copyright (C) Metaswitch Networks 2017
+# If license terms are provided to you in a COPYING file in the root directory
+# of the source code repository by which you are accessing this code, then
+# the license outlined in that COPYING file applies to your use.
+# Otherwise no rights are granted except for those provided to you by
+# Metaswitch Networks in a separate written agreement.
+
+FILENAME=/etc/clearwater/default_ifcs.xml
+SCHEMA=/usr/share/clearwater/clearwater-config-manager/scripts/config_validation/default_ifcs_schema.xsd
+
+. /usr/share/clearwater/utils/check-root-permissions 1
+
+[ $# -le 1 ] || { echo "Usage: validate_default_ifcs_xml [file to validate] (defaults to $FILENAME if not specified)" >&2 ; exit 2 ; }
+[ -z "$1" ] || FILENAME=$1
+
+# Check that the file to validate is present.
+if [[ ! -f $FILENAME ]]
+then
+    echo "No configuration file at $FILENAME, unable to validate"
+    exit 2
+fi
+
+xmllint --format --pretty 1 --load-trace --debug --schema $SCHEMA $FILENAME 2> /tmp/validate-default-ifcs-xml.stderr.$$ > /tmp/validate-default-ifcs-xml.stdout.$$
+rc=$?
+
+# Check the return code and log if appropriate.
+if [ $rc != 0 ] ; then
+  echo The Default IFC file is invalid                    >&2
+  cat /tmp/validate-default-ifcs-xml.stderr.$$            >&2
+else
+  echo The Default IFC file is valid
+fi
+
+rm -f /tmp/validate-default-ifcs-xml.stderr.$$ /tmp/validate-default-ifcs-xml.stdout.$$
+
+exit $rc
