--- clearwater_config_manager/scripts/validate_shared_ifcs_config
@@ -1,61 +1,0 @@
-#!/bin/bash
-
-# Project Clearwater - IMS in the Cloud
-# Copyright (C) 2017  Metaswitch Networks Ltd
-#
-# This program is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or (at your
-# option) any later version, along with the "Special Exception" for use of
-# the program along with SSL, set forth below. This program is distributed
-# in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-# without even the implied warranty of MERCHANTABILITY or FITNESS FOR
-# A PARTICULAR PURPOSE.  See the GNU General Public License for more
-# details. You should have received a copy of the GNU General Public
-# License along with this program.  If not, see
-# <http://www.gnu.org/licenses/>.
-#
-# The author can be reached by email at clearwater@metaswitch.com or by
-# post at Metaswitch Networks Ltd, 100 Church St, Enfield EN2 6BQ, UK
-#
-# Special Exception
-# Metaswitch Networks Ltd  grants you permission to copy, modify,
-# propagate, and distribute a work formed by combining OpenSSL with The
-# Software, or a work derivative of such a combination, even if such
-# copying, modification, propagation, or distribution would otherwise
-# violate the terms of the GPL. You must comply with the GPL in all
-# respects for all of the code used other than OpenSSL.
-# "OpenSSL" means OpenSSL toolkit software distributed by the OpenSSL
-# Project and licensed under the OpenSSL Licenses, or a work based on such
-# software and licensed under the OpenSSL Licenses.
-# "OpenSSL Licenses" means the OpenSSL License and Original SSLeay License
-# under which the OpenSSL Project distributes the OpenSSL toolkit software,
-# as those licenses appear in the file LICENSE-OPENSSL.
-
-FILENAME=/etc/clearwater/shared_ifcs.xml
-SCHEMA=/usr/share/clearwater/clearwater-config-manager/scripts/config_validation/shared_ifcs_schema.xsd
-
-. /usr/share/clearwater/utils/check-root-permissions 1
-
-[ $# -le 1 ] || { echo "Usage: validate_shared_ifcs_config [file to validate] (defaults to $FILENAME if not specified)" >&2 ; exit 2 ; }
-[ -z "$1" ] || FILENAME=$1
-
-# Check that the file to validate is present.
-if [[ ! -f $FILENAME ]]
-then
-    echo "No configuration file at $FILENAME, unable to validate"
-    exit 2
-fi
-
-xmllint --format --pretty 1 --load-trace --debug --schema $SCHEMA $FILENAME 2> /tmp/validate-shared-ifcs-config.stderr.$$ > /tmp/validate-shared-ifcs-config.stdout.$$
-rc=$?
-
-# Check the return code and log if appropriate.
-if [ $rc != 0 ] ; then
-  echo The Shared IFC file is invalid                       >&2
-  cat /tmp/validate-shared-ifcs-config.stderr.$$            >&2
-  cat /tmp/validate-shared-ifcs-config.stdout.$$            >&2
-fi
-rm -f /tmp/validate-shared-ifcs-config.stderr.$$ /tmp/validate-shared-ifcs-config.stdout.$$
-
-exit $rc
