+++ clearwater_config_manager/scripts/validate_shared_ifcs_xml
@@ -1,0 +1,37 @@
+#!/bin/bash
+
+# Copyright (C) Metaswitch Networks 2017
+# If license terms are provided to you in a COPYING file in the root directory
+# of the source code repository by which you are accessing this code, then
+# the license outlined in that COPYING file applies to your use.
+# Otherwise no rights are granted except for those provided to you by
+# Metaswitch Networks in a separate written agreement.
+
+FILENAME=/etc/clearwater/shared_ifcs.xml
+SCHEMA=/usr/share/clearwater/clearwater-config-manager/scripts/config_validation/shared_ifcs_schema.xsd
+
+. /usr/share/clearwater/utils/check-root-permissions 1
+
+[ $# -le 1 ] || { echo "Usage: validate_shared_ifcs_config [file to validate] (defaults to $FILENAME if not specified)" >&2 ; exit 2 ; }
+[ -z "$1" ] || FILENAME=$1
+
+# Check that the file to validate is present.
+if [[ ! -f $FILENAME ]]
+then
+    echo "No configuration file at $FILENAME, unable to validate"
+    exit 2
+fi
+
+xmllint --format --pretty 1 --load-trace --debug --schema $SCHEMA $FILENAME 2> /tmp/validate-shared-ifcs-xml.stderr.$$ > /tmp/validate-shared-ifcs-xml.stdout.$$
+rc=$?
+
+# Check the return code and log if appropriate.
+if [ $rc != 0 ] ; then
+  echo The Shared IFC file is invalid                    >&2
+  cat /tmp/validate-shared-ifcs-xml.stderr.$$            >&2
+else
+  echo The Shared IFC file is valid
+fi
+rm -f /tmp/validate-shared-ifcs-xml.stderr.$$ /tmp/validate-shared-ifcs-xml.stdout.$$
+
+exit $rc
